Parsed the binding file in an unstructured way to avoid reflection
